Handle null and undefined values
`transformRawValue` expects a string as input
So when it receives a number, for example, it's transforming '1' to 1
But null and undefined are also valid values on react native styles (e.g. `width: null` is used sometimes)
Currently if we pass `transformRawValue('width', null)` it will crash and if we pass `transformRawValue('width', 'null')` it will return `{ width: 'null' }` which is not a valid property.